backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
python-multipart
python-pptx
PyPDF2
python-docx
openpyxl
